requirements.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.11
django-appconf==1.0.3
django-compressor==2.3
django-crispy-forms==1.8.1
django-debug-toolbar==2.2
django-formtools==2.2
django-haystack==2.8.1
django-import-export==2.0.2
django-mdeditor==0.1.16
django-redis==4.11.0
django-reversion==3.0.7
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
docutils==0.15.2
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
idna==2.8
importlib-metadata==0.23
jdcal==1.4.1
jieba==0.39
kombu==4.6.5
MarkupPy==1.14
mistune==0.8.4
more-itertools==7.2.0
mysqlclient==1.4.4
odfpy==1.4.1
openpyxl==3.0.3
Pillow==6.2.1
Pygments==2.4.2
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
PyYAML==5.3
rcssmin==1.0.6
redis==3.3.11
requests==2.22.0
rjsmin==1.1.0
six==1.12.0
sqlparse==0.3.0
tablib==1.1.0
Unidecode==1.1.1
urllib3==1.25.6
vine==1.3.0
Whoosh==2.7.4
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
